Integrate Get data to API
