+++ go.mod
@@ -4,21 +4,13 @@ go 1.14
 
 require (
 	github.com/caarlos0/env v3.5.0+incompatible
-	github.com/go-playground/universal-translator v0.17.0 // indirect
-	github.com/go-playground/validator v9.31.0+incompatible
 	github.com/go-playground/validator/v10 v10.2.0
 	github.com/google/wire v0.4.0
 	github.com/kljensen/snowball v0.6.0
-	github.com/kr/pretty v0.2.0 // indirect
-	github.com/labstack/echo v3.3.10+incompatible
 	github.com/labstack/echo/v4 v4.1.16
-	github.com/leodido/go-urn v1.2.0 // indirect
-	github.com/pkg/errors v0.8.1
 	github.com/rs/zerolog v1.18.0
 	github.com/stretchr/testify v1.5.1
 	github.com/xlab/closer v0.0.0-20190328110542-03326addb7c2
 	github.com/xujiajun/nutsdb v0.5.0
-	github.com/ziflex/lecho/v2 v2.0.0 // indirect
 	github.com/zoomio/stopwords v0.5.0
-	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
 )
